+++ src/main/java/com/deltasf/createpropulsion/registries/PropulsionFluids.java
@@ -25,15 +25,21 @@ import net.minecraftforge.client.extensions.common.IClientFluidTypeExtensions;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.FluidType;
 import net.minecraftforge.fluids.ForgeFlowingFluid;
+import net.minecraftforge.fml.DistExecutor;
+
 import com.simibubi.create.foundation.utility.Color;
 
 public class PropulsionFluids {
     public static final CreateRegistrate REGISTRATE = CreatePropulsion.registrate();
     public static void register() {} //Loads this class
 
-    public static final FluidEntry<ForgeFlowingFluid.Flowing> TURPENTINE = REGISTRATE.standardFluid("turpentine", 
-        SolidRenderedPlaceableFluidType.create(0xd69e49, () -> 1.0f / 4.0f, 0xffd69e49))
-        .renderType(() -> RenderType.translucent())
+    private static final Supplier<FluidTypeFactory> TURPENTINE_TYPE_FACTORY = DistExecutor.unsafeRunForDist(
+        () -> PropulsionFluidsClient::getTurpentineTypeFactory,
+        () -> PropulsionFluids::createGenericFactory
+    );
+
+    public static final FluidEntry<ForgeFlowingFluid.Flowing> TURPENTINE = REGISTRATE.standardFluid("turpentine", TURPENTINE_TYPE_FACTORY.get())
+        .renderType(getSidedRenderType())
         .lang("Turpentine")
         .properties(p -> p.viscosity(1000).density(500))
         .fluidProperties(p -> p.levelDecreasePerBlock(1)
@@ -41,120 +47,17 @@ public class PropulsionFluids {
             .slopeFindDistance(3)
             .explosionResistance(100f))
         .register();
+
+    //Helpers
     
-
-    public static abstract class TintedFluidType extends FluidType {
-        private ResourceLocation stillTexture;
-        private ResourceLocation flowingTexture;
-
-        public TintedFluidType(Properties properties, ResourceLocation stillTexture, ResourceLocation flowingTexture) {
-            super(properties);
-            this.stillTexture = stillTexture;
-            this.flowingTexture = flowingTexture;
-        }
-
-        @Override
-        public void initializeClient(Consumer<IClientFluidTypeExtensions> consumer) {
-            consumer.accept(new IClientFluidTypeExtensions() {
-
-                @Override
-                public ResourceLocation getStillTexture() {
-                    return stillTexture;
-                }
-
-                @Override
-                public ResourceLocation getFlowingTexture() {
-                    return flowingTexture;
-                }
-
-                @Override
-                public int getTintColor(FluidStack stack) {
-                    return TintedFluidType.this.getTintColor(stack);
-                }
-
-                @Override
-                public int getTintColor(FluidState state, BlockAndTintGetter getter, BlockPos pos) {
-                    return TintedFluidType.this.getTintColor(state, getter, pos);
-                }
-
-                @Override
-                public @NotNull Vector3f modifyFogColor(Camera camera, float partialTick, ClientLevel level,
-                                                        int renderDistance, float darkenWorldAmount, Vector3f fluidFogColor) {
-                    Vector3f customFogColor = TintedFluidType.this.getCustomFogColor();
-                    return customFogColor == null ? fluidFogColor : customFogColor;
-                }
-
-                @Override
-                public void modifyFogRender(Camera camera, FogMode mode, float renderDistance, float partialTick,
-                                            float nearDistance, float farDistance, FogShape shape) {
-                    float modifier = TintedFluidType.this.getFogDistanceModifier();
-                    float baseWaterFog = 96.0f;
-                    if (modifier != 1f) {
-                        RenderSystem.setShaderFogShape(FogShape.CYLINDER);
-                        RenderSystem.setShaderFogStart(-8);
-                        RenderSystem.setShaderFogEnd(baseWaterFog * modifier);
-                    }
-                }
-
-            });
-        }
-
-        protected abstract int getTintColor(FluidStack stack);
-
-        protected abstract int getTintColor(FluidState state, BlockAndTintGetter getter, BlockPos pos);
-
-        protected Vector3f getCustomFogColor() {
-            return null;
-        }
-
-        protected float getFogDistanceModifier() {
-            return 1f;
-        }
-
+    private static Supplier<RenderType> getSidedRenderType() {
+        return DistExecutor.unsafeRunForDist(
+                () -> PropulsionFluidsClient::getTurpentineRenderType,
+                () -> () -> null
+        );
     }
-
-    private static class SolidRenderedPlaceableFluidType extends TintedFluidType {
-
-        private Vector3f fogColor;
-        private Supplier<Float> fogDistance;
-        private int tintColor;
-
-        public static FluidTypeFactory create(int fogColor, Supplier<Float> fogDistance, int tintColor) {
-            return (p, s, f) -> {
-                s = new ResourceLocation("block/water_still");
-                f = new ResourceLocation("block/water_flow");
-                SolidRenderedPlaceableFluidType fluidType = new SolidRenderedPlaceableFluidType(p, s, f);
-                fluidType.fogColor = new Color(fogColor, false).asVectorF();
-                fluidType.fogDistance = fogDistance;
-                fluidType.tintColor = tintColor;
-                return fluidType;
-            };
-        }
-
-        private SolidRenderedPlaceableFluidType(Properties properties, ResourceLocation stillTexture,
-                                                ResourceLocation flowingTexture) {
-            super(properties, stillTexture, flowingTexture);
-        }
-
-        @Override
-        protected int getTintColor(FluidStack stack) {
-            return tintColor;
-        }
-
-        @Override
-        public int getTintColor(FluidState state, BlockAndTintGetter world, BlockPos pos) {
-            return tintColor;
-        }
-
-        @Override
-        protected Vector3f getCustomFogColor() {
-            return fogColor;
-        }
-
-        @Override
-        protected float getFogDistanceModifier() {
-            return fogDistance.get();
-        }
-
+    
+    private static Supplier<FluidTypeFactory> createGenericFactory() {
+        return () -> (properties, stillTexture, flowingTexture) -> new FluidType(properties);
     }
 }
