+++ src/main/java/com/deltasf/createpropulsion/registries/PropulsionFluidsClient.java
@@ -1,0 +1,153 @@
+package com.deltasf.createpropulsion.registries;
+
+import java.util.function.Consumer;
+import java.util.function.Supplier;
+
+import org.jetbrains.annotations.NotNull;
+import org.joml.Vector3f;
+
+import com.jozufozu.flywheel.util.Color;
+import com.mojang.blaze3d.shaders.FogShape;
+import com.mojang.blaze3d.systems.RenderSystem;
+import com.tterrag.registrate.builders.FluidBuilder.FluidTypeFactory;
+
+import net.minecraft.client.Camera;
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.FogRenderer.FogMode;
+import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.BlockAndTintGetter;
+import net.minecraft.world.level.material.FluidState;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraftforge.client.extensions.common.IClientFluidTypeExtensions;
+import net.minecraftforge.fluids.FluidStack;
+import net.minecraftforge.fluids.FluidType;
+
+@OnlyIn(Dist.CLIENT)
+public class PropulsionFluidsClient {
+    
+    public static Supplier<FluidTypeFactory> getTurpentineTypeFactory() {
+        return () -> SolidRenderedPlaceableFluidType.create(0xd69e49, () -> 1.0f / 4.0f, 0xffd69e49);
+    }
+
+    public static Supplier<RenderType> getTurpentineRenderType() {
+        return () -> RenderType.translucent();
+    }
+
+    public static abstract class TintedFluidType extends FluidType {
+        private ResourceLocation stillTexture;
+        private ResourceLocation flowingTexture;
+
+        public TintedFluidType(Properties properties, ResourceLocation stillTexture, ResourceLocation flowingTexture) {
+            super(properties);
+            this.stillTexture = stillTexture;
+            this.flowingTexture = flowingTexture;
+        }
+
+        @Override
+        public void initializeClient(Consumer<IClientFluidTypeExtensions> consumer) {
+            consumer.accept(new IClientFluidTypeExtensions() {
+
+                @Override
+                public ResourceLocation getStillTexture() {
+                    return stillTexture;
+                }
+
+                @Override
+                public ResourceLocation getFlowingTexture() {
+                    return flowingTexture;
+                }
+
+                @Override
+                public int getTintColor(FluidStack stack) {
+                    return TintedFluidType.this.getTintColor(stack);
+                }
+
+                @Override
+                public int getTintColor(FluidState state, BlockAndTintGetter getter, BlockPos pos) {
+                    return TintedFluidType.this.getTintColor(state, getter, pos);
+                }
+
+                @Override
+                public @NotNull Vector3f modifyFogColor(Camera camera, float partialTick, ClientLevel level,
+                                                        int renderDistance, float darkenWorldAmount, Vector3f fluidFogColor) {
+                    Vector3f customFogColor = TintedFluidType.this.getCustomFogColor();
+                    return customFogColor == null ? fluidFogColor : customFogColor;
+                }
+
+                @Override
+                public void modifyFogRender(Camera camera, FogMode mode, float renderDistance, float partialTick,
+                                            float nearDistance, float farDistance, FogShape shape) {
+                    float modifier = TintedFluidType.this.getFogDistanceModifier();
+                    float baseWaterFog = 96.0f;
+                    if (modifier != 1f) {
+                        RenderSystem.setShaderFogShape(FogShape.CYLINDER);
+                        RenderSystem.setShaderFogStart(-8);
+                        RenderSystem.setShaderFogEnd(baseWaterFog * modifier);
+                    }
+                }
+
+            });
+        }
+
+        protected abstract int getTintColor(FluidStack stack);
+
+        protected abstract int getTintColor(FluidState state, BlockAndTintGetter getter, BlockPos pos);
+
+        protected Vector3f getCustomFogColor() {
+            return null;
+        }
+
+        protected float getFogDistanceModifier() {
+            return 1f;
+        }
+
+    }
+
+    private static class SolidRenderedPlaceableFluidType extends TintedFluidType {
+
+        private Vector3f fogColor;
+        private Supplier<Float> fogDistance;
+        private int tintColor;
+
+        public static FluidTypeFactory create(int fogColor, Supplier<Float> fogDistance, int tintColor) {
+            return (p, s, f) -> {
+                s = new ResourceLocation("block/water_still");
+                f = new ResourceLocation("block/water_flow");
+                SolidRenderedPlaceableFluidType fluidType = new SolidRenderedPlaceableFluidType(p, s, f);
+                fluidType.fogColor = new Color(fogColor, false).asVectorF();
+                fluidType.fogDistance = fogDistance;
+                fluidType.tintColor = tintColor;
+                return fluidType;
+            };
+        }
+
+        private SolidRenderedPlaceableFluidType(Properties properties, ResourceLocation stillTexture,
+                                                ResourceLocation flowingTexture) {
+            super(properties, stillTexture, flowingTexture);
+        }
+
+        @Override
+        protected int getTintColor(FluidStack stack) {
+            return tintColor;
+        }
+
+        @Override
+        public int getTintColor(FluidState state, BlockAndTintGetter world, BlockPos pos) {
+            return tintColor;
+        }
+
+        @Override
+        protected Vector3f getCustomFogColor() {
+            return fogColor;
+        }
+
+        @Override
+        protected float getFogDistanceModifier() {
+            return fogDistance.get();
+        }
+
+    }
+}
